Add support for caching with CasADi SXFunctions
With the support for picklable SXFunctions in CasADi 3.4.0, we can now
cache with a pickle object instead of needing code generation. With
pickling, it is possible to later still call ca.substitute() and
.expand() to simplify the graph, which can give significant performance
gains.

Code generation can still be useful for models that would not benefit
much from any later simplifcations to the graph (e.g. when there is
nothing to simplify anyway), so we will still support that with the
"codegen" option.

Note that currently caching implies expansion of the MX graph, as
pickling of MXFunctions is not yet supported.

Closes #56